Refactor: 마이페이지 user info 섹션 컴포넌트 분리
