+++ frontend/src/pages/mypage/MyPage.tsx
@@ -1,9 +1,10 @@
 import { useSuspenseQuery } from '@tanstack/react-query'
 
+import { Separator } from '@/components/ui'
 import { profileQuries } from '@/service/api'
 import { useMyInfoStore } from '@/store'
 
-import { UserInfoSection, UserSocialInfoSection } from './_components'
+import { ProfileImage, ProfileInfo, UserSocialInfoSection } from './_components'
 
 export default function MyPage() {
   const { userId } = useMyInfoStore((state) => state.myInfo)
@@ -14,7 +15,12 @@ export default function MyPage() {
 
   return (
     <div className="flex w-full flex-col items-center justify-center pb-20">
-      <UserInfoSection profile={profile} userId={userId} />
+      <section className="flex w-full flex-col md:flex-row md:space-x-8">
+        <ProfileImage profile={profile} userId={userId} />
+        <Separator orientation="vertical" className="hidden md:flex" />
+        <Separator className="my-10 md:hidden" />
+        <ProfileInfo profile={profile} />
+      </section>
       <UserSocialInfoSection profile={profile} userId={userId} />
     </div>
   )
