+++ frontend/src/pages/mypage/_components/index.ts
@@ -1,2 +1,3 @@
-export * from './user-info'
+export * from './profile-image'
 export * from './user-social-info'
+export * from './profile-info'
