+++ frontend/src/pages/mypage/_components/profile-info/index.ts
@@ -1,0 +1,1 @@
+export { ProfileInfo } from './ProfileInfo'
