--- frontend/src/pages/mypage/_components/user-info/index.ts
@@ -1,1 +1,0 @@
-export { UserInfoSection } from './UserInfoSection'
